+++ src/linker.rs
@@ -1,50 +1,15 @@
-use crate::{
-    compiler::{ExePath, ObjPath},
-    error::{BackEndError, BackEndErrorInfo, BackEndErrorKind, ErrorInfo},
-};
-use std::{io, process::Command};
+use crate::compiler::{ExePath, ObjPath};
+use std::process::Command;
 
 #[derive(Clone, Copy, Debug)]
 pub struct Linker;
 
 impl Linker {
-    pub fn link(obj_path: &ObjPath, exe_path: &ExePath) -> Result<(), BackEndError<ErrorKind>> {
+    pub fn link(obj_path: &ObjPath, exe_path: &ExePath) -> Command {
         let obj_path_str = obj_path.inner.to_str().unwrap();
         let exe_path_str = exe_path.inner.to_str().unwrap();
-        let args = [&obj_path_str, "-o", &exe_path_str];
-
-        match Command::new("ld").args(args).output() {
-            Ok(ld_out) if !ld_out.status.success() => {
-                Err(BackEndError { kind: ErrorKind::Failed { output: ld_out.stderr } })
-            }
-            Ok(_) => Ok(()),
-            Err(err) => Err(BackEndError { kind: ErrorKind::CouldNotCreateProcess { err } }),
-        }
+        let mut linker_command = Command::new("ld");
+        let _ = linker_command.args([&obj_path_str, "-o", &exe_path_str]);
+        linker_command
     }
 }
-
-#[derive(Debug)]
-pub enum ErrorKind {
-    CouldNotCreateProcess { err: io::Error },
-    Failed { output: Vec<u8> },
-}
-
-impl ErrorInfo for ErrorKind {
-    type Info = BackEndErrorInfo;
-
-    fn info(&self) -> Self::Info {
-        let (msg, cause) = match self {
-            Self::CouldNotCreateProcess { err } => (
-                "could not create linker process".into(),
-                format!("{err} ({kind})", kind = err.kind()).into(),
-            ),
-            Self::Failed { output } => {
-                ("linker failed".into(), String::from_utf8_lossy(output).into_owned().into())
-            }
-        };
-
-        Self::Info { msg, cause }
-    }
-}
-
-impl BackEndErrorKind for ErrorKind {}