+++ CHANGELOG.md
@@ -45,6 +45,9 @@ Changes to the ABI would **not** be considered breaking since no stable ABI is d
 - Improved code generation
 - Consistency of enum's field order
 - Started running examples as tests to check if output is correct
+- `run` command now returns the exit code of the executed program
+- `compile` command now returns the exit code of the assembler and linker in the case of a failure
+- Compilation steps now return the relative `std::process::Command` to allow for finer control
 - **BREAKING**: `char` type renamed to `ascii`
 - **BREAKING**: Runtime crashes now print to `stderr`
 - **BREAKING**: `Compiler::compile` now returns an `Artifacts` struct